Authorizer refactored
