+++ api-security/src/main/java/gov/ca/cwds/security/permission/AbacPermission.java
@@ -3,6 +3,7 @@ package gov.ca.cwds.security.permission;
 import com.google.inject.Key;
 import com.google.inject.name.Names;
 import gov.ca.cwds.security.authorizer.Authorizer;
+import gov.ca.cwds.security.authorizer.BaseAuthorizer;
 import gov.ca.cwds.security.module.SecurityModule;
 import org.apache.shiro.authz.Permission;
