+++ api-security/src/test/java/gov/ca/cwds/security/authorizer/CaseAuthorizer.java
@@ -6,7 +6,7 @@ import gov.ca.cwds.testapp.domain.Case;
  * Created by dmitry.rudenko on 10/6/2017.
  */
 
-public class CaseAuthorizer extends Authorizer<Case, Long> {
+public class CaseAuthorizer extends BaseAuthorizer<Case, Long> {
   public boolean checkId(Long id) {
     return id == 1L;
   }
